Add highlighting to user-selected squares
Implemented a feature to visually highlight squares in the Sudoku Grid that have been interactively chosen by the user. Also added a new 'highlight' prop to SquareChoice and updated the SquareDisplay component to effectively pass this prop. Adjusted the CSS rules in square.css to change the color of highlighted squares and updated the SquareChoice stories in Storybook to include the new 'highlight' prop.